Add Login + Route auth

+++ routes/web.php
@@ -21,28 +21,34 @@ use App\Http\Controllers\ListingController;
 Route::get('/', [ListingController::class, 'index']);
 
 //Show Create Form
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 //Show Store new Listing
 Route::post('/listings', [ListingController::class, 'store']);
 
 //Show Store new Listing
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 // Update Listing
 Route::put('/listings/{listing}', [ListingController::class, 'update']);
 
 // Delete Listing
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
-//Show single Listing
+// Show single Listing
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
-//Show User Registration
-Route::get('/register', [UserController::class, 'create']); 
+// Show User Registration
+Route::get('/register', [UserController::class, 'create'])->middleware('guest'); 
 
-//Create new User
+// Create new User
 Route::post('/users', [UserController::class, 'store']);
 
-//Logout User
-Route::post('/logout', [UserController::class, 'logout']);
+// Logout User
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
+
+// Show Login Form
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+
+// Log In User
+Route::post('/users/authenticade', [UserController::class, 'authenticade']);
